create_fucntion_database_check
